==> src/07 Estimacion Diferencia Salarial.Rmd <==
---
title: "Untitled"
author: "Lorena Rodriguez Chamorro"
date: "28 de septiembre de 2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, include=FALSE}
require("tidyverse")
require("dplyr")
require("ggplot2")
library(rattle)
library(rpart, quietly=TRUE)
```

```{r, include=FALSE}
#Cargo la tabla del conjunto de datos Femenino
load("../dat/fem.data.rds")

fem.data %>%
  select(SALANUAL, SALESTIMADO, FACTOTAL) %>%
  summarize(Real= (sum(SALANUAL * FACTOTAL) / sum(FACTOTAL)), Estimado= (sum(SALESTIMADO * FACTOTAL) / sum(FACTOTAL)),
            diferencia = Estimado-Real, DifPorcentual=((Estimado-Real)/Real)*100)

fem.data$diferencia <- fem.data$SALESTIMADO-fem.data$SALANUAL
fem.data$DifPorcentual <- ((fem.data$SALESTIMADO-fem.data$SALANUAL)/fem.data$SALANUAL)*100
#Mas de un 10% de diferencia con el dato real entonces existe diferencia 
fem.data$ExisteDiferencia <-ifelse(fem.data$DifPorcentual>10,1,0) 

```

## ANÁLISIS DIFERENCIA SALARIAL

Una vez estudiada la brecha salarial a través de la sobreestimación del salario femenino partiendo del modelo creado con el salario masculino, se va a realizar un modelo que calcule la probabilidad de que exista dicha diferencia salarial. Esto nos va a permitir encontrar combinaciones de condiciones laborales que tengan mayor o menor probabilidad de que exista diferencia entre el salario masculino y el femenino.

Para ello, vamos a crear una nueva variable que tenga valor 1 si existe una diferencia entre el salario estimado de las mujeres y el real superior al 10%, tendrá valor 0 en caso contrario.

Por lo tanto, si estudiamos la totalidad del conjunto de datos de las mujeres (64.383 asalariadas), la variable ExisteDiferencia nos queda distribuida de la siguiente forma:

```{r, echo=FALSE}
table(fem.data$ExisteDiferencia)
```

Para analizar que condiciones son las que tienen mayor probabilidad de que exista diferencia salarial vamos a crear un modelo de árbol con rpart que nos clasifique la variable ExisteDiferencia. 

```{r, include=FALSE}
#04. Seleccion de variables
fem.arbol <- fem.data %>%
  select(-c(ORDENCCC, ORDENTRA))%>%
  select(-c(ANOANTI,MESANTI,FIJODISM,FIJODISD,DRELABM,SIESPM1,DSIESPM1,SIESPM2,DSIESPM2,SALBASE,
            EXTRAORM,PHEXTRA,COMSAL,COMSALTT,IRPFMES,COTIZA,BASE,DRELABAM,DRELABAD,
            SALBRUTO,GEXTRA,VESP,FACTOTAL,DIASANO,DES_CNACE,DES_TCNO,TIPOJOR)) %>%
  select(-c(DifPorcentual,SALANUAL,SALESTIMADO,diferencia))

fem.arbol$CNO1 <- ifelse(fem.arbol$CNO1=="A0", "A0.Directores/Gerentes",
                  ifelse(fem.arbol$CNO1=="B0", "B0.Tec/Prof cient Salud/Enseñ",
                  ifelse(fem.arbol$CNO1=="C0", "C0.Otros Tec/Prof cientificos",
                  ifelse(fem.arbol$CNO1=="D0", "D0.Tec/Prof apoyo",
                  ifelse(fem.arbol$CNO1=="E0", "E0.Emp Oficina no Publico",
                  ifelse(fem.arbol$CNO1=="F0", "F0.Emp Oficina si Publico",
                  ifelse(fem.arbol$CNO1=="G0", "G0.Trab serv restauracion",
                  ifelse(fem.arbol$CNO1=="H0", "H0.Trab serv salud",
                  ifelse(fem.arbol$CNO1=="I0", "I0.Trab serv seguridad",
                  ifelse(fem.arbol$CNO1=="J0", "J0.Trab cuali agricolas",             
                  ifelse(fem.arbol$CNO1=="K0", "K0.Trab cuali construccion",
                  ifelse(fem.arbol$CNO1=="L0", "L0.Trab cuali industrias",
                  ifelse(fem.arbol$CNO1=="M0", "M0.Operadores instalaciones",
                  ifelse(fem.arbol$CNO1=="N0", "N0.Conductores maquinaria",
                  ifelse(fem.arbol$CNO1=="O0", "O0.Trab Serv no cualificados",
                  ifelse(fem.arbol$CNO1=="P0", "P0.Peones agricultura/pesca",
                  ifelse(fem.arbol$CNO1=="Q0", "Q0.Ocupaciones militares",
                  "")))))))))))))))))


#05. Entrenamos el modelo 
tree.model <- rpart(ExisteDiferencia ~ .,
                   data=fem.arbol,
                   method="class",
                   parms=list(split="information"),
                   control=rpart.control(cp=0.0020,
                                         usesurrogate=0, 
                                         maxsurrogate=0),
                   model=TRUE)

```

```{r, echo=FALSE}
#09. Pintamos el arbol
fancyRpartPlot(tree.model, main="Arbol de decision")
```


Las reglas que nos devuelve este modelo son las siguientes:

```{r, echo=FALSE}
#08. Sacamos las reglas
asRules(tree.model)
```

###CONCLUSIONES

Podemos concluir que la combinación del tipo de actividad, la ocupación, la zona geográfica, la jornada anual pactada, los días en situación especial y los meses de antigüedad nos va a determinar si existe mayor o menor probabilidad de que existan diferencias salariales.
